Resolve import collisions (#1)
* feat: Resolve import collisions

We resolve import collisions by first collecting the required schemas
from the proto messages. Then we sort these schemas by the local package
prefix which allows us to resolve the local message references first and
then resolve the import messages. If there is a collision we prefix the
imported package to the schema message otherwise the message name is
left as is.

* fix: comment