fix: handle int64 as string
context:
- https://github.com/grpc-ecosystem/grpc-gateway/issues/438
- https://github.com/golang/protobuf/issues/1414
- https://github.com/OpenAPITools/openapi-generator-cli

Originally, this was changed so that int64 fields would be written to
the openapi spec as integer types which is supported by the OpenAPI
spec. However, jsonpb and other related projects have chosen to enforce
that int64 fields are written as strings. This is because JavaScript
represents integers as 32 bit numbers and if you pass it a value that is
outside of that range it will overflow. Since we are using jsonpb to
transform the protos to json, we need to make sure that the generated
schema is consistent with the jsonpb behavior.